Isolamento CSS do componente CustomDrawer
