+++ src/components/CustomDrawer/index.js
@@ -1,0 +1,3 @@
+import CustomDrawer from './CustomDrawer';
+
+export default CustomDrawer;
